+++ package.json
@@ -6,11 +6,15 @@
   "scripts": {
     "eslint": "eslint src/**/*.js",
     "test": "npm run eslint",
+    "build": "webpack",
     "start": "webpack-dev-server --open"
   },
   "devDependencies": {
     "eslint": "6.1.0",
-    "eslint-config-htmlacademy": "0.5.2"
+    "eslint-config-htmlacademy": "0.5.2",
+    "webpack": "4.41.2",
+    "webpack-cli": "3.3.10",
+    "webpack-dev-server": "3.9.0"
   },
   "engines": {
     "node": "10"
