+++ src/components/utils.js
@@ -1,0 +1,7 @@
+export {setUpperCase};
+const setUpperCase = (str) => {
+  if (!str) {
+    return str;
+  }
+  return str[0].toUpperCase() + str.slice(1);
+};
